+++ benchmarks/all_graphs_on_n_verts.py
@@ -10,6 +10,7 @@ sys.path.append(parent_dir)
 import msr
 
 SAVE_FIGS = True
+FIG_DIR = f"figs/"
 QUIET = False
 
 
@@ -20,11 +21,8 @@ def main(n: int):
     accurate the bounds are is generated.
     """
 
-    # directory where graphs are saved
-    graph_dir = f"../msr/graph/saved/n{n}/"
-
     # compute bounds on all graphs
-    bounds_and_ids = msr.msr_batch_from_directory(path=graph_dir, quiet=QUIET)
+    bounds_and_ids = msr.msr_batch_from_directory(quiet=QUIET, num_verts=n)
 
     # find troublemakers
     not_tight = 0
@@ -49,12 +47,11 @@ def main(n: int):
         )
 
     # save figures of troublemakers
-    fig_dir = f"figs/n{n}/"
     if SAVE_FIGS and len(ids) > 0:
         print("saving figures...")
         for id in tqdm(ids):
-            graph_file = graph_dir + f"k{id}" + ".json"
-            image_file = fig_dir + f"k{id}" + ".png"
+            graph_file = msr.graph.SAVED_GRAPH_DIR + f"{id}.graph"
+            image_file = FIG_DIR + f"k{id}.png"
             msr.graph.draw_graph(
                 G=msr.graph.load_graph(graph_file),
                 embedding="min_entropy",