+++ msr/__init__.py
@@ -1,11 +1,10 @@
 from . import graph
 from .msr_batch import msr_batch, msr_batch_from_directory
-from .msr_bounds import msr, msr_bounds
+from .msr_bounds import msr_bounds
 from .msr_sdp import msr_sdp_upper_bound
 
 __all__ = [
     "graph",
-    "msr",
     "msr_batch",
     "msr_batch_from_directory",
     "msr_bounds",