+++ msr/graph/generate.py
@@ -1,21 +1,25 @@
-import os
 from itertools import permutations
-from math import ceil, factorial, log10
+from math import factorial
 from multiprocessing import Pool
+from typing import Optional
 
 import tqdm
 
-from .file_io import save_graph
+from .file_io import SAVED_GRAPH_DIR, save_graphs
 from .graph import graph
 
 
-def generate_and_save_all_graphs_on_n_vertices(n: int, path: str) -> None:
+def generate_and_save_all_graphs_on_n_vertices(
+    n: int, path: Optional[str] = None
+) -> None:
     """Generates and saves all graphs on n vertices"""
-    nx_graphs = generate_all_graphs_on_n_vertices(n)
-    save_graphs(nx_graphs, path)
+    if path is None:
+        path = SAVED_GRAPH_DIR + f"n{n}"
+    graphs = generate_all_graphs_on_n_vertices(n)
+    save_graphs(graphs, path)
 
 
-def generate_all_graphs_on_n_vertices(n: int) -> set[graph]:
+def generate_all_graphs_on_n_vertices(n: int) -> list[graph]:
     """
     Generates all graphs on n vertices by constructing all graphs isomorphic to
     G and testing if any of these graphs have been seen.
@@ -23,8 +27,7 @@ def generate_all_graphs_on_n_vertices(n: int) -> set[graph]:
     Idea from https://stackoverflow.com/questions/71597789/generate-all-digraphs-of-a-given-size-up-to-isomorphism
     """
 
-    n_choose_2 = n * (n - 1) // 2
-    num_candidates = 2**n_choose_2
+    num_candidates = 2 ** (n * (n - 1) // 2)
 
     found_hashes: set[int] = set()
     encountered: list[bool] = list([False for _ in range(num_candidates)])
@@ -34,23 +37,23 @@ def generate_all_graphs_on_n_vertices(n: int) -> set[graph]:
     for k in tqdm.tqdm(range(num_candidates)):
         if encountered[k]:
             continue
-        G = construct_graph_from_hash(k, n, n_choose_2)
+        G = graph(num_verts=n)
+        G.build_from_hash(k)
         if G.num_edges() < n - 1:
             continue
         is_not_new, seen = is_not_new_graph(G, found_hashes)
         for t in seen:
             encountered[t] = True
         if not is_not_new:
-            found_hashes.add(hash(G))
+            found_hashes.add(k)
 
     # reconstruct graphs from hashes
-    graphs = set()
+    graphs = []
     for k in found_hashes:
-        # G = construct_graph_from_hash(k, n, n_choose_2)
         G = graph(num_verts=n)
         G.build_from_hash(k)
         if G.is_connected():
-            graphs.add(G)
+            graphs.append(G)
 
     # report number of graphs
     print(f"Number of connected graphs on {n} vertices: {len(graphs)}")
@@ -89,43 +92,10 @@ def is_not_new_graph_worker(
     return hash(H), hash(H) in found_hashes
 
 
-def construct_graph_from_hash(k: int, n: int, n_choose_2: int) -> graph:
-    """Constructs a graph on n vertices from an edge hash k."""
-
-    # convert k to binary, and pad with zeros to the left
-    binary = bin(k)[2:].zfill(n_choose_2)
-
-    # convert binary to a list of edges in a graph G
-    G = graph(num_verts=n)
-    for i in range(n - 1):
-        for j in range(i + 1, n):
-            ij = j - 1 + (i * (2 * n - 3 - i)) // 2
-            if binary[ij] == "1":
-                G.add_edge(i, j)
-
-    return G
-
-
-def save_graphs(graphs: set[graph], path: str) -> None:
-    """Saves graphs to disk."""
-
-    # create directory if none exists
-    if not os.path.exists(path):
-        os.makedirs(path)
-
-    # save graphs
-    print(f"Saving graphs to {path}")
-    for G in tqdm.tqdm(graphs):
-        filename = f"{path}/k{hash(G)}.json"
-        save_graph(G, filename)
-
-
 def num_graphs_on_n_verts(n: int) -> int:
     """
     OEIS A001349: Number of connected graphs with n nodes.
     """
-    if n > 20:
-        raise ValueError("n must be <= 20")
     a = [
         1,
         1,
@@ -148,4 +118,7 @@ def num_graphs_on_n_verts(n: int) -> int:
         1787331725248899088890200576580,
         24636021429399867655322650759681644,
     ]
+    a_len = len(a)
+    if n > a_len - 1:
+        print("WARNING: number of graphs on n vertices not found in OEIS table")
     return a[n]