+++ msr/msr_batch.py
@@ -2,18 +2,20 @@ import multiprocessing
 
 from tqdm import tqdm
 
-from .graph.file_io import files_in_directory, load_graph
+from .graph.file_io import SAVED_GRAPH_DIR, files_in_directory, load_graph
 from .graph.graph import graph
 from .msr_bounds import msr_bounds
 
 
 def msr_batch_from_directory(
-    path: str, quiet: bool = False
-) -> list[tuple[int, int, int]]:
+    path: str = SAVED_GRAPH_DIR,
+    quiet: bool = False,
+    num_verts: int = 0,
+) -> list[tuple[int, int, str]]:
     """
     Computes the MSR bounds for graphs in a directory with multiprocessing.
     """
-    filenames = files_in_directory(path)
+    filenames = files_in_directory(path, num_verts)
     if quiet:
         return _msr_batch_from_directory_quiet(filenames)
     else:
@@ -22,7 +24,7 @@ def msr_batch_from_directory(
 
 def _msr_batch_from_directory_loud(
     filenames: list[str],
-) -> list[tuple[int, int, int]]:
+) -> list[tuple[int, int, str]]:
     """Uses tqdm to show progress."""
     with multiprocessing.Pool() as pool:
         return list(
@@ -35,7 +37,7 @@ def _msr_batch_from_directory_loud(
 
 def _msr_batch_from_directory_quiet(
     filenames: list[str],
-) -> list[tuple[int, int, int]]:
+) -> list[tuple[int, int, str]]:
     """Does not use tqdm to show progress."""
     with multiprocessing.Pool() as pool:
         return list(
@@ -43,7 +45,7 @@ def _msr_batch_from_directory_quiet(
         )
 
 
-def _msr_bounds_with_id_from_file(filename: str) -> tuple[int, int, int]:
+def _msr_bounds_with_id_from_file(filename: str) -> tuple[int, int, str]:
     """Helper function for msr_batch_from_directory."""
     G = load_graph(filename)
     return _msr_bounds_with_id(G)
@@ -51,7 +53,7 @@ def _msr_bounds_with_id_from_file(filename: str) -> tuple[int, int, int]:
 
 def msr_batch(
     graphs: list[graph], quiet: bool = False
-) -> list[tuple[int, int, int]]:
+) -> list[tuple[int, int, str]]:
     """Computes the MSR bounds for a batch of graphs with multiprocessing."""
     if quiet:
         return _msr_batch_quiet(graphs)
@@ -59,7 +61,7 @@ def msr_batch(
         return _msr_batch_loud(graphs)
 
 
-def _msr_batch_loud(graphs: list[graph]) -> list[tuple[int, int, int]]:
+def _msr_batch_loud(graphs: list[graph]) -> list[tuple[int, int, str]]:
     """Uses tqdm to show progress."""
     N = len(graphs)
     with multiprocessing.Pool() as pool:
@@ -68,13 +70,13 @@ def _msr_batch_loud(graphs: list[graph]) -> list[tuple[int, int, int]]:
         )
 
 
-def _msr_batch_quiet(graphs: list[graph]) -> list[tuple[int, int, int]]:
+def _msr_batch_quiet(graphs: list[graph]) -> list[tuple[int, int, str]]:
     """Does not use tqdm to show progress."""
     with multiprocessing.Pool() as pool:
         return list(pool.imap_unordered(_msr_bounds_with_id, graphs))
 
 
-def _msr_bounds_with_id(G: graph) -> tuple[int, int, int]:
+def _msr_bounds_with_id(G: graph) -> tuple[int, int, str]:
     """Helper function for msr_batch."""
     d_lo, d_hi = msr_bounds(G)
-    return d_lo, d_hi, hash(G)
+    return d_lo, d_hi, G.id()